Lock both the bucket and the element when inserting a new part
Although simply locking the bucket suffices for mutual exclusion between
a pair of concurrent insertions, compiler and CPU reordering can in theory
cause problems for concurrent lookups, which might see pre-initialization
garbage.  This commit therefore makes insertions acquire the lock not only
for the bucket, but also for the element being inserted.  This properly
synchronizes with lookups as well as with concurrent insertions.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>